+++ pdfminer/layout.py
@@ -51,10 +51,15 @@ class LAParams:
         of a text matters when determining the order of text boxes. The value
         should be within the range of -1.0 (only horizontal position
         matters) to +1.0 (only vertical position matters).
+    :param cell_margin: (float) An additional distance allowing a line
+        to be counted as splitting an object when split_table mode enabled.
     :param detect_vertical: If vertical text should be considered during
         layout analysis
     :param all_texts: If layout analysis should be performed on text in
         figures.
+    :param split_tables: (bool) Enable parsing logic to have textboxes
+        and textlines to be bounded LTRect, LTLine, and LTCurve objects.
+        This causes the parser to respect boundary lines drawn a table.
     """
 
     def __init__(self,
@@ -80,10 +85,11 @@ class LAParams:
 
     def __repr__(self):
         return '<LAParams: char_margin=%.1f, line_margin=%.1f, ' \
-               'word_margin=%.1f, cell_margin=%.1f, ' \
-               'all_texts=%r, split_tables=%r>' % \
+               'word_margin=%.1f, cell_margin=%.1f, boxes_flow=%.1f, ' \
+               'detect_vertical=%r, all_texts=%r, split_tables=%r>' % \
                (self.char_margin, self.line_margin, self.word_margin,
-                self.cell_margin, self.all_texts, self.split_tables)
+                self.cell_margin, self.boxes_flow, self.detect_vertical,
+                self.all_texts, self.split_tables)
 
 
 class LTItem: